+++ src/libs/radcore/src/radmemorymonitor/memmonitorclient.hpp
@@ -33,7 +33,7 @@
 #include <radmemorymonitor.hpp>
 #include <radmemory.hpp>
 #include <radthread.hpp>
-#include "..\radprotocols\memorymonitorprotocol.hpp"
+#include "../radprotocols/memorymonitorprotocol.hpp"
 
 //=============================================================================
 // Forward Class Declarations
@@ -263,7 +263,7 @@ protected:
     //
     // internal client information declaration
     //
-	void DeclareMemSpaceInfo( radMemorySpace memSpace, unsigned int startAddr, unsigned int size );
+	void DeclareMemSpaceInfo( radMemorySpace memSpace, uintptr_t startAddr, unsigned int size );
 
     void DeclareMemSpaceInfo( );
 
@@ -275,7 +275,7 @@ protected:
 
     void SendSuspendState( );
     
-    void SendStackUsageData( radMemorySpace memSpace, unsigned int uAddress );
+    void SendStackUsageData( radMemorySpace memSpace, uintptr_t uAddress );
 
     // used for DeclareSection()
     void DecalareAllocationFromTraverse( radMemorySpace memorySpace, IRadMemoryTraverse* traverse );
@@ -286,9 +286,9 @@ protected:
 
     void NameSection( void* address, radMemorySpace memorySpace, const char * name );
 
-    void SendMemoryBlock128Byte( radMemorySpace memorySpace, unsigned int memStartPos, unsigned int memLength );
+    void SendMemoryBlock128Byte( radMemorySpace memorySpace, uintptr_t memStartPos, unsigned int memLength );
 
-    void SendMemoryBlock( radMemorySpace memorySpace, unsigned int memStartPos, unsigned int memLength );
+    void SendMemoryBlock( radMemorySpace memorySpace, uintptr_t memStartPos, unsigned int memLength );
 
     void SendObjectRefount( radMemorySpace memorySpace, void * pObjectPtr, unsigned int * pRefCount );
 
@@ -304,7 +304,7 @@ protected:
     // resume the console operation, in pair with Suspend( )
     void Resume( );
 
-    bool IsMemoryBlockValid( radMemorySpace memorySpace, unsigned int memStartPos, unsigned int memLength );
+    bool IsMemoryBlockValid( radMemorySpace memorySpace, uintptr_t memStartPos, unsigned int memLength );
 
     bool RemoveCorrespondingAddRefInBuffer( void* pObject, void* pReference, radMemorySpace memorySpaceObject );
